# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "JobPostingMetadataUpdate": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "company_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "job_title": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "recruitment_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecruitmentStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "title": "JobPostingMetadataUpdate",
        "type": "object"
      },
      "JobPostingResponse": {
        "properties": {
          "active_analysis_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Analysis Run Id"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "collected_at": {
            "format": "date-time",
            "title": "Collected At",
            "type": "string"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          },
          "recruitment_stage": {
            "$ref": "#/components/schemas/RecruitmentStage"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/JobPostingStatus"
          },
          "target_role_id": {
            "format": "uuid",
            "title": "Target Role Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "active_analysis_run_id",
          "target_role_id",
          "company_name",
          "job_title",
          "recruitment_stage",
          "city",
          "source_url",
          "original_text",
          "status",
          "collected_at",
          "created_at",
          "updated_at"
        ],
        "title": "JobPostingResponse",
        "type": "object"
      },
      "JobPostingStatus": {
        "enum": [
          "draft",
          "queued",
          "extracting",
          "analyzing",
          "review_required",
          "confirmed",
          "failed"
        ],
        "title": "JobPostingStatus",
        "type": "string"
      },
...
      "ProblemDetails": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemValidationError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemValidationError": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "pointer"
        ],
        "title": "ProblemValidationError",
        "type": "object"
      },
...
      "RecruitmentStage": {
        "enum": [
          "daily_internship",
          "summer_internship",
          "winter_internship",
          "autumn_recruitment",
          "spring_recruitment",
          "other"
        ],
        "title": "RecruitmentStage",
        "type": "string"
      },
      "RequirementExplicitness": {
        "enum": [
          "explicit",
          "implicit",
          "uncertain"
        ],
        "title": "RequirementExplicitness",
        "type": "string"
      },
      "RequirementItemCreate": {
        "properties": {
          "explicitness": {
            "$ref": "#/components/schemas/RequirementExplicitness"
          },
          "normalized_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Normalized Name",
            "type": "string"
          },
          "original_text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Original Text",
            "type": "string"
          },
          "requirement_type": {
            "$ref": "#/components/schemas/RequirementType"
          }
        },
        "required": [
          "original_text",
          "normalized_name",
          "requirement_type",
          "explicitness"
        ],
        "title": "RequirementItemCreate",
        "type": "object"
      },
      "RequirementItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RequirementItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "job_status": {
            "$ref": "#/components/schemas/JobPostingStatus"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "job_status"
        ],
        "title": "RequirementItemListResponse",
        "type": "object"
      },
      "RequirementItemResponse": {
        "properties": {
          "analysis_run_id": {
            "format": "uuid",
            "title": "Analysis Run Id",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "explicitness": {
            "$ref": "#/components/schemas/RequirementExplicitness"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "normalized_name": {
            "title": "Normalized Name",
            "type": "string"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          },
          "requirement_type": {
            "$ref": "#/components/schemas/RequirementType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_confirmed": {
            "title": "User Confirmed",
            "type": "boolean"
          },
          "user_modified": {
            "title": "User Modified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "analysis_run_id",
          "original_text",
          "normalized_name",
          "requirement_type",
          "explicitness",
          "confidence",
          "user_confirmed",
          "user_modified",
          "created_at",
          "updated_at"
        ],
        "title": "RequirementItemResponse",
        "type": "object"
      },
      "RequirementItemUpdate": {
        "properties": {
          "explicitness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RequirementExplicitness"
              },
              {
                "type": "null"
              }
            ]
          },
          "normalized_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Name"
          },
          "original_text": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text"
          },
          "requirement_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RequirementType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RequirementItemUpdate",
        "type": "object"
      },
...
      "RequirementType": {
        "enum": [
          "eligibility",
          "core_competency",
          "experience",
          "preferred",
          "uncertain"
        ],
        "title": "RequirementType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "岗位门槛分析系统 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_posting_endpoint_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Delete Job Posting Endpoint",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_posting_endpoint_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPostingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Get Job Posting Endpoint",
        "tags": [
          "jobs"
        ]
      },
      "patch": {
        "operationId": "update_job_posting_endpoint_api_v1_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPostingMetadataUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPostingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Update Job Posting Endpoint",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}/confirm-requirements": {
      "post": {
        "operationId": "confirm_requirements_endpoint_api_v1_jobs__job_id__confirm_requirements_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPostingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Confirm Requirements Endpoint",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/requirements": {
      "get": {
        "operationId": "list_requirements_endpoint_api_v1_jobs__job_id__requirements_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementItemListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "List Requirements Endpoint",
        "tags": [
          "requirements"
        ]
      },
      "post": {
        "operationId": "create_requirement_endpoint_api_v1_jobs__job_id__requirements_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Create Requirement Endpoint",
        "tags": [
          "requirements"
        ]
      }
    },
...
    "/api/v1/requirements/{requirement_id}": {
      "delete": {
        "operationId": "delete_requirement_endpoint_api_v1_requirements__requirement_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Delete Requirement Endpoint",
        "tags": [
          "requirements"
        ]
      },
      "get": {
        "operationId": "get_requirement_endpoint_api_v1_requirements__requirement_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Get Requirement Endpoint",
        "tags": [
          "requirements"
        ]
      },
      "patch": {
        "operationId": "update_requirement_endpoint_api_v1_requirements__requirement_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Update Requirement Endpoint",
        "tags": [
          "requirements"
        ]
      }
    },
...
  }
}